Validate that requests don't have extra parameters
Request validation is enhanced to verify requests only include query or
formData parameters that are specified in the spec.

This validation does not occur for header or path parameters.  This is
because most applications probably prefer to ignore extra headers and
a request with extra path parameters would point to a different
endpoint.