Remove six package from the dev test suite
The url for the six package repository was wrong. Now it is changed
to the correct one, but tests still fail. We need to investigate,
but probably some change in other packages are causing the problem.
Since it is not really connected to Connexion code, I have decided
to remove it from the test suite.

Signed-off-by: Rafael Caricio <rafael.caricio@zalando.de>