Fix errors when returning flask Response objects
This will allow you to return flask Response objects such as
```return jsonify(['a']), 201, {'new_header': 'header_value'```.
It also fixes errors that occur when returning a Response or
ResponseContainer with ```validate_responses=True```.

This is a breaking change, ResponseContainer __init__ response kwarg
has been removed, and replaced with a ```from_response``` method. I
think this is better, as there isn't any ambiguity between what happens
if you use a response with data in it (or various other situations like
that), and takes special care to make sure the data can actually be
validated without modifying or breaking parts of the original flask
Response object. That said, if we are against making a breaking change
here, I could modify this to work with the single __init__ method and
response kwarg without much work.