Removed duplicated calls of get_response method of the Api classes (#809)
On Operation decorators lifecycle the api.get_response method were always called twice.
Because the response validation, we need to cast the response to the ConnexionResponse type.
But the response casting could be made on the response validation decorator, not always. So, changes were made to cast the response correctly.
The BeginOfRequestLifecycleDecorator class was removed because it only do the second call of api.get_response method. This second call only occurs if the response validation is needed.
For a more logical naming, the EndOfRequestLifecycleDecorator class was renamed to RequestResponseDecorator.
Fixes #664