Skip plain/text content_type test in aiohttp
The test cannot pass without revisiting how aiohttp_api uses _cast_body.
Instead of using the response content_type, it is casting the body based
on the openapi spec defined mediatype, which does not match in this case
when trying to return application/problem+json.

This also leaves an open question: How should problems be serialized if
the spec defines a format other than application/problem+json, or
application/json, etc? Here, a simple plain/text triggers the issue.

Ultimate reason for skipping the test (recorded in the test):

  aiohttp_api.get_connexion_response uses _cast_body to stringify the
  dict directly instead of using json.dumps. This differs from flask
  usage, where there is no _cast_body.