Standardize data serialization
The old aiohttp_api._cast_body did not jsonify when mimetype was None.
The old flask_api._jsonify_data jsonified data, or raised a TypeError
if the data could not be jsonified, even when mimetype was not None
and when mimetype was not a JSON mimetype.

This unifies the behavior so that only JSON or None mimetypes can trigger
jsonification, and only JSON mimetypes will raise a TypeError if the data
is not jsonifiable. Non-JSON mimetypes and non-jsonifiable data with None
mimetype are stringified.

This means that the data can serialize itself when
  - mimetype is None and data is not jsonifiable
  - mimetype is not a JSON mimetype

This removes the flask_api version of _jsonify_data. This should be
backwards compatible because:

For the non-jsonifiable data, when mimetype is None or a non-json type,
using str() instead of throwing TypeError is an improvement. Could anyone
depend on throwing a TypeError in this case? It would just throw 500
errors, but this makes that more useful. So it is a "new feature".

The only gotcha is when mimetype is not None and is not JSON, but the
data is jsonifiable, we should NOT serialize it as JSON, but today we
are for flask but not for aiohttp. Instead flask_api should just cast it
as a string (like aiohttp does) assuming that if someone wanted a
non-json type, they will handle serialization, possibly by subclassing
dict with their own__str__ implementation or something. This is probably
a bugfix (and therefore ok to introduce now) because jsonifying data
with a non-json mimetype with flask does not make sense, even if the
data is jsonifiable.