Bugfix/missing infofuncs (#1375)
* Add test for verify_security with missing auth funcs

This situation occurs when a x-infoFunc is missing in the openapi
document. Then this security scheme is skipped and the auth_funcs
is an empty list if there are no other security schemes.

* Fix verify_security bug for empty list of auth_funcs

When the auth_funcs argument is the empty list, an AttributeError
is raised because the token_info is None and does not have a
'get' attribute. This leads to a 500 error.