Fix OpenAPI parameters containing other parameters (#1523)
* Fix OpenAPI parameters containing other parameters

OpenAPI parameters can be unspecified and sometimes contain other
parameters. The current behavior is to assume it's a bracket parameter
and nest the containing parameter within the contained parameter, which
breaks the schema as the original parameter now seems to contain a
nested object.

We can avert this by checking for the presence of a '[' in the
parameter.

* Trigger Github workflow

Co-authored-by: Ricardo Piro-Rael <rpiro-rael@ironox.com>