Split parsing from validation completely (#1934)
During the refactoring for Connexion 3, we deduplicated all
`coerce_type` calls during the refactoring and moved them into the
`uri_parser` so it is done in a single place. When looking into #1931
however, I noticed we are still using the `uri_parser` from more places
than we should.

This PR centralizes the parsing further into the `ConnexionRequest`
class. During validation, we now instantiate a `ConnexionRequest`
instead of a Starlette `Request`, and leverage it instead of calling the
`uri_parser` directly.

I split the parsing and validation in the tests so they can be tested in
isolation.