Custom auth types in security_map (#1940)
Allow definition of new auth types (not just schemes) to be passed in to
security_map
This is particularly motivated by wanting to support OpenIDConnect auth
type, but was running into an error as it couldn't find the `scheme` key
(as it's not permitted in the schema by the openAPI validator).

The
[docs](https://connexion.readthedocs.io/en/stable/security.html#security)
imply that you should be able to implement a custom scheme for this,
instead I would get a
[`KeyError`](https://gist.github.com/anna-intellegens/0161d967de923030a1cc9f822160a13a).

Changes proposed in this pull request:

- Update `parse_security_schemes` to allow the detection of auth types,
not just schemes.

---------

Co-authored-by: Ruwann <ruwanlambrichts@gmail.com>